chore: 🚚 move error pages to appropriate location
They were erroneously in `/components/` when they should have either been in `/pages` or `/components/pages/global` respectively.
