+++ apps/ui/src/pages/AppRouter.tsx
@@ -20,7 +20,6 @@
 import { Navigate, Route, Routes } from 'react-router';
 
 import PageLayout from '@/components/layouts/PageLayout';
-import NotFound from '@/components/pages/NotFound';
 import ProtectedRoute from '@/components/ProtectedRoute';
 import ApplicationViewer from '@/pages/applications';
 import AccessAgreement from '@/pages/applications/sections/access';
@@ -37,9 +36,11 @@ import DashboardPage from '@/pages/dashboard';
 import HomePage from '@/pages/index';
 import LoginRedirect from '@/pages/login/redirect';
 import ManageApplicationsPage from '@/pages/manage/applications';
+import NotFound from '@/pages/NotFound';
 import InstitutionalRepLogin from '@/pages/review';
 import { ApplicationContextProvider } from '@/providers/context/application/ApplicationContextProvider';
 import { SectionRoutes, SectionRoutesValues } from '@pcgl-daco/validation';
+import LoginError from './login/error';
 
 export interface ApplicationSectionRouteTypes {
 	route: SectionRoutesValues;
@@ -105,6 +106,7 @@ function AppRouter() {
 		<Routes>
 			<Route element={<PageLayout />}>
 				<Route path="login/redirect" element={<LoginRedirect />} />
+				<Route path="login/error/*" element={<LoginError />} />
 
 				<Route index element={<HomePage />} />
 				<Route
