+++ apps/ui/src/pages/login/error/index.tsx
@@ -1,0 +1,7 @@
+import ErrorPage from '@/components/pages/global/ErrorPage';
+
+const LoginError = () => {
+	return <ErrorPage error={{ message: 'Skill Issue', error: "Can't log you in." }} loading={false}></ErrorPage>;
+};
+
+export default LoginError;
